Use the go4.org/mem package to reduce allocations.
For many of the common operations in the scanner and parser, we can get rid of
redundant byte-to-string copies by using the mem.RO type, which provides a view
of the underlying data.

Move the internals of the Quote and Unquote functions to an internal package
that uses the mem.RO type, and rework the exported versions as wrappers.

Benchmarks:

  name                       old time/op    new time/op    delta
  Scanner/JTree/Scanner-10     1.47ms ± 0%    1.41ms ± 0%   -3.89%
  Scanner/JTree/Stream-10      1.52ms ± 0%    1.46ms ± 0%   -4.39%
  Scanner/JTree/ParseAST-10    2.22ms ± 0%    2.17ms ± 0%   -2.35%

  name                       old alloc/op   new alloc/op   delta
  Scanner/JTree/Scanner-10     7.70kB ± 0%    4.69kB ± 0%  -39.09%
  Scanner/JTree/Stream-10      7.93kB ± 0%    4.82kB ± 0%  -39.15%
  Scanner/JTree/ParseAST-10     727kB ± 0%     723kB ± 0%   -0.43%

  name                       old allocs/op  new allocs/op  delta
  Scanner/JTree/Scanner-10        580 ± 0%         6 ± 0%  -98.97%
  Scanner/JTree/Stream-10         583 ± 0%         8 ± 0%  -98.63%
  Scanner/JTree/ParseAST-10     20.8k ± 0%     20.2k ± 0%   -2.77%