ast: avoid overloading the String method
Switch ast.Stringer back to ast.Texter, so that the choice of making a value
convertible to a JSON string is more deliberate. String methods are used as
magic by the fmt package, so you could wind up doing this by accident.

At the same time, add String methods to all the various Value types, so that
they render nicely in formatted output.