tq: rework environments and binding
The handling of bindings was asymmetric and confusing. The Get and Set queries
would mutate the environment in-place, while Let would bind names in scope.
Let was also clunky to write, and the "nice" way of using it with inline
bindings triggers lint warnings because of the unkeyed initializers.

Summary:

- extend the Query interface to return a new environment.
- each environment now contains (only) one binding.
- remove the Let and Set constructs.
- add As to do bindings.
- rework Func to return environments too.