ast: reduce allocation on the parse stack
Prior to this change, the parser would preallocate objects and arrays, and
eagerly populate their elements as they were reduced. This turned out to be a
lot worse for memory use than pushing everything onto the primary stack.

Instead, introduce marker values for incomplete object and array values, and do
a single allocation for each of them in a single reduction when the end marker
is found. This substantially reduces both time and allocations.

  name                       old time/op    new time/op    delta
  Scanner/JTree/ParseAST-10    2.17ms ± 0%    2.00ms ± 0%   -7.99%

  name                       old alloc/op   new alloc/op   delta
  Scanner/JTree/ParseAST-10     621kB ± 0%     515kB ± 0%  -17.06%

  name                       old allocs/op  new allocs/op  delta
  Scanner/JTree/ParseAST-10     20.2k ± 0%     15.6k ± 0%  -22.84%