scanner: fix a corner case when lexing block comments (#17)
When seeing the block comment /***/, the scanner looks after the opening /* for
the end of the comment. Upon reaching a "*", it checks whether the next thing
is "/" which would end the comment.

If not, it resumes searching, but previously (in that case) the failing rune
was accumulated into the token. In this example, that means it has now consumed
the second star, which means the reader incorrectly treats the following "/" as
part of the comment body.

Instead, we need to unread the failing rune so that it gets picked up by the
next search pass. Do this, and update the tests to backstop it.