jwcc: make ToValue usable for compound values
Previously, any argument not already a jwcc.Value would be wrapped in a Datum.
But ast.ToValue already passes through arrays and objects unmodified, so we
really should do the same for jwcc.ToValue too.

Use the unwrapping in Decorate, which already know how to handle a compound
ast.Value, and then we only need to check specifically for jwcc.Value up front
(as we do not want to down-convert those).