jwcc: format commented member values more charitably
When the value of an object member has its own before comments, bump the whole
value in one additional level of indentation, e.g.

    "foo":
      // the value
      "bar",

Moreover, drop the inlining behaviour for block comments, so that instead of

    "foo":
      /* the value */ bar,

we now format it as above:

    "foo":
      /* the value */
      "bar",

which seems nicer to read. This leaves the inline check unused, and maybe we
don't need it anymore, but I left it in for now.